+++ src/jax_fdm/visualization/plotters/__init__.py
@@ -1,4 +1,9 @@
 from .loss_plotter import *  # noqa F403
+from .network import *  # noqa F403
+from .plotter import *  # noqa F403
+from .register import register_artists
 
+
+register_artists()
 
 __all__ = [name for name in dir() if not name.startswith('_')]